requirements.txt
fastapi
uvicorn
pandas
jinja2
python-multipart
google-cloud-storage
google-cloud-bigquery
